Adds person class and phase module
