+++ apprentice_teacher.rb
@@ -1,21 +1,11 @@
-class ApprenticeTeacher
-  attr_reader :age, :salary, :phase, :target_raise
-  attr_accessor :name
+require_relative 'person'
+
+class ApprenticeTeacher < Person
+  attr_reader :performance_rating, :target_raise
+
 
   def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
     @target_raise = 800
-    @phase = 3
-  end
-
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
   end
 
   def teach_stuff
@@ -24,15 +14,6 @@ class ApprenticeTeacher
     response += "*drops crazy knowledge bomb* "
     response += "... You're welcome."
     response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
@@ -51,3 +32,6 @@ class ApprenticeTeacher
     puts "Whoa. I know ruby-fu"
   end
 end
+
+# Same on this one, due to difference in reaction of different output I cannot store it in overall class. 
+# I want to keep Person class clean.
