+++ person.rb
@@ -1,0 +1,29 @@
+class Person
+  attr_reader :age, :phase, :salary, :performance_rating, :target_raise
+  attr_accessor :name
+
+  def initialize(options={})
+    @phase = phase
+    @age = options[:age]
+    @name = options[:name]
+  end
+
+  def offer_high_five
+    "High five!"
+  end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+   def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+end
