+++ phase.rb
@@ -1,0 +1,25 @@
+require_relative 'person'
+
+module Phase
+  def initialize(options = {})
+  end
+
+  def set_phase(num)
+   response = ""
+   if num == @phase
+     response += "I'm doing phase #{@phase} again because "
+     response += "I put my learning first. I'm gonna rock it!"
+   else
+     response = "Oooh, phase #{num}. I hope I'm ready!"
+   end
+    @phase = num
+    response
+  end
+
+  def learn_stuff
+    response = ""
+    response += "WHOA! I've never thought of it quite like that before. "
+    response += "Now I feel like a genius!"
+    response
+  end
+end
